Move MessageDestructionTimeoutValue to new file
